+++ core/api/users.py
@@ -18,7 +18,7 @@ except ImportError:
     except ImportError:
         import slave  # type: ignore
 
-PLUGIN_VERSION = "0.7.37"
+PLUGIN_VERSION = "0.7.39"
 
 
 def _extract_param(request, key, default=""):